bao/requirements.txt
joblib==1.2.0
numpy==1.23.5
psycopg2-binary==2.9.5
scikit-learn==1.1.3
scipy==1.9.3
threadpoolctl==3.1.0
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.13.0+cpu
typing_extensions==4.4.0
